refactor: move app into src rather than inside components
